+++ FactorioProfiles/src/Data.cs
@@ -23,6 +23,11 @@ namespace FactorioProfiles
 		// file. If a database file exists, this value should be read from there. 
 		private static string defaultProfileSavePath = @"%APPDATA%\Powershell\FactorioProfiles\Profiles";
 
+		// The current database version.
+		// ⚠ This should only be used when creating a new database file. If an existing database file exists,
+		// perform the migration process until you reach this version.
+		private static string databaseVersion = "0.1.1";
+
 		private static XmlDocument OpenFile()
 		{
 			var document = new XmlDocument();
@@ -37,6 +42,7 @@ namespace FactorioProfiles
 				// Load an existing configuration file if one exists.
 				// TODO: Validate against a schema.
 				document.Load(fullDatabasePath);
+				MigrateData(document);
 			}
 			else
 			{
@@ -49,7 +55,7 @@ namespace FactorioProfiles
 				// Create a 'Data' body element to contain all of the data nodes.
 				// Give it a version attribute, to allow for data migration if the data structure ever changes.
 				var body = document.CreateElement("Data");
-				body.SetAttribute("Version", "0.1.0");
+				body.SetAttribute("Version", databaseVersion);
 				document.AppendChild(body);
 
 				// Create a 'Config' element to contain configuration data which is related to the module
@@ -83,6 +89,10 @@ namespace FactorioProfiles
 				config.AppendChild(defaultSharing);
 				body.AppendChild(config);
 
+				// Create an element to contain the information about which is the currently active profile.
+				var activeProfile = document.CreateElement("ActiveProfile");
+				body.AppendChild(activeProfile);
+
 				// Create an element to contain all of the 'Profile' class data from serialization.
 				var profiles = document.CreateElement("Profiles");
 				body.AppendChild(profiles);
@@ -98,6 +108,13 @@ namespace FactorioProfiles
 				databasePath);
 
 			document.Save(fullDatabasePath);
+		}
+
+		public static String GetGlobalProfilePath()
+		{
+			return System.IO.Path.Combine(
+				System.Environment.GetFolderPath(System.Environment.SpecialFolder.ApplicationData),
+				globalProfilePath);
 		}
 
 		public static String GetNewProfileSavePath()
@@ -191,6 +208,36 @@ namespace FactorioProfiles
 			}
 
 			return null;
+		}
+
+		public static String GetActiveProfile()
+		{
+			// Load in the document.
+			var document = OpenFile();
+
+			// Select the 'ActiveProfile' node.
+			var node = document.DocumentElement.SelectSingleNode("/Data/ActiveProfile");
+
+			// If there is a value return that, otherwise return N/A.
+			// This can happen if the database has just been migrated but the 'Switch' cmdlet hasn't ran yet,
+			// if which case this will hold an empty value.
+			if (String.IsNullOrWhiteSpace(node.InnerText))
+			{
+				return "N/A";
+			}
+			return node.InnerText;
+		}
+
+		public static void SetActiveProfile(Profile profile)
+		{
+			// Load in the document.
+			var document = OpenFile();
+
+			// Select the 'ActiveProfile' node value.
+			document.DocumentElement.SelectSingleNode("/Data/ActiveProfile").InnerText = profile.Name;
+
+			// Save the changes to disk.
+			CloseFile(document);
 		}
 
 		public static void Add(Profile profile)
@@ -326,6 +373,30 @@ namespace FactorioProfiles
 
 			// Save the changes to disk.
 			CloseFile(document);
+		}
+
+		private static void MigrateData(XmlDocument document)
+		{
+			var version = document.SelectSingleNode("/Data").Attributes.GetNamedItem("Version").Value;
+
+			// Loop, migrating the database to each consecutive version until it's migrated to the latest.
+			while (!version.Equals(databaseVersion, StringComparison.OrdinalIgnoreCase))
+			{
+				if (version.Equals("0.1.0", StringComparison.OrdinalIgnoreCase))
+				{
+					// 0.1.0 -> 0.1.1
+					var activeProfile = document.CreateElement("ActiveProfile");
+					document.SelectSingleNode("/Data").InsertBefore(
+						activeProfile,
+						document.SelectSingleNode("/Data/Profiles")
+					);
+
+					document.SelectSingleNode("/Data").Attributes.GetNamedItem("Version").Value = "0.1.1";
+				}
+
+				// Update the version number for the loop check.
+				version = document.SelectSingleNode("/Data").Attributes.GetNamedItem("Version").Value;
+			}
 		}
 	}
 }